Add hex entry option
Adds an input to enable setting the active color by entering a hex
value. When a hex is entered, the other elements adjust accordingly and
the name input clears unless the hex is an exact match for a named
color.

Also started to make the color shown on initial load nicer.